整理 js, 分離一些 function
