+++ homeworks/week9/hw1/css/style.css
@@ -14,10 +14,6 @@ p {
   margin: 0;
 }
 
-iframe {
-  margin: 10px auto;
-}
-
 .navbar {
   display: flex;
   height: 50px;
