+++ homeworks/week9/hw1/utlis.php
@@ -4,6 +4,7 @@ require_once("conn.php");
 // markdown 與 XSS 過濾
 require_once("Parsedown.php");
 $parsedown = new Parsedown();
+$parsedown->setMarkupEscaped(true);
 // 一些實用 function
 // 產生 token
